add AddBook page
