+++ Client/src/components/AddBook.jsx
@@ -1,0 +1,71 @@
+// src/AddBook.js
+import React, { useState } from 'react';
+import { createBook } from '../Servers/bookAPI';
+import Swal from 'sweetalert2';
+
+const AddBook = () => {
+  const [newBook, setNewBook] = useState({ name: '', author: '', genre: '', limit_age: '', status: '' });
+
+  const handleCreateBook = async () => {
+    Swal.fire({
+      title: "Do you want to add the book?",
+      showDenyButton: true,
+      showCancelButton: true,
+      confirmButtonText: "Save",
+      denyButtonText: `Don't save`
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        try {
+          await createBook(newBook);
+          setNewBook({ name: '', author: '', genre: '', limit_age: '', status: '' });
+          Swal.fire("Saved!", "", "success");
+        } catch (error) {
+          console.error(error);
+        }
+      } else if (result.isDenied) {
+        Swal.fire("Changes are not saved", "", "info");
+      }
+    });
+  };
+
+  return (
+    <div>
+      <h1>Add Book</h1>
+      <div>
+        <input
+          type="text"
+          placeholder="Name"
+          value={newBook.name}
+          onChange={(e) => setNewBook({ ...newBook, name: e.target.value })}
+        />
+        <input
+          type="text"
+          placeholder="Author"
+          value={newBook.author}
+          onChange={(e) => setNewBook({ ...newBook, author: e.target.value })}
+        />
+        <input
+          type="text"
+          placeholder="Genre"
+          value={newBook.genre}
+          onChange={(e) => setNewBook({ ...newBook, genre: e.target.value })}
+        />
+        <input
+          type="text"
+          placeholder="Limit Age"
+          value={newBook.limit_age}
+          onChange={(e) => setNewBook({ ...newBook, limit_age: e.target.value })}
+        />
+        <input
+          type="number"
+          placeholder="Status"
+          value={newBook.status}
+          onChange={(e) => setNewBook({ ...newBook, status: e.target.value })}
+        />
+        <button onClick={handleCreateBook}>Create Book</button>
+      </div>
+    </div>
+  );
+};
+
+export default AddBook;
