+++ Client/src/components/Book.jsx
@@ -1,29 +1,12 @@
-// // src/Books.js
 // import React, { useState, useEffect } from 'react';
 // import { getAllBooks, createBook, deleteBook, updateBook } from '../Servers/bookAPI';
 // import { Link } from 'react-router-dom';
+// import Swal from 'sweetalert2';
 
 // const Book = () => {
 //   const [books, setBooks] = useState([]);
 //   const [newBook, setNewBook] = useState({ name: '', author: '', genre: '', limit_age: '', status: '' });
 //   const [updateBookId, setUpdateBookId] = useState(null);
-
-
-
-//   Swal.fire({
-//     title: "Do you want to add the book?",
-//     showDenyButton: true,
-//     showCancelButton: true,
-//     confirmButtonText: "Save",
-//     denyButtonText: `Don't save`
-//   }).then((result) => {
-//     /* Read more about isConfirmed, isDenied below */
-//     if (result.isConfirmed) {
-//       Swal.fire("Saved!", "", "success");
-//     } else if (result.isDenied) {
-//       Swal.fire("Changes are not saved", "", "info");
-//     }
-//   });
 
 //   useEffect(() => {
 //     fetchBooks();
@@ -39,33 +22,72 @@
 //   };
 
 //   const handleCreateBook = async () => {
-//     try {
-//       await createBook(newBook);
-//       fetchBooks();
-//       setNewBook({ name: '', author: '', genre: '', limit_age: '', status: '' });
-//     } catch (error) {
-//       console.error(error);
-//     }
+//     Swal.fire({
+//       title: "Do you want to add the book?",
+//       showDenyButton: true,
+//       showCancelButton: true,
+//       confirmButtonText: "Save",
+//       denyButtonText: `Don't save`
+//     }).then(async (result) => {
+//       if (result.isConfirmed) {
+//         try {
+//           await createBook(newBook);
+//           fetchBooks();
+//           setNewBook({ name: '', author: '', genre: '', limit_age: '', status: '' });
+//           Swal.fire("Saved!", "", "success");
+//         } catch (error) {
+//           console.error(error);
+//         }
+//       } else if (result.isDenied) {
+//         Swal.fire("Changes are not saved", "", "info");
+//       }
+//     });
 //   };
 
 //   const handleDeleteBook = async (id) => {
-//     try {
-//       await deleteBook(id);
-//       fetchBooks();
-//     } catch (error) {
-//       console.error(error);
-//     }
+//     Swal.fire({
+//       title: "Do you want to delete the book?",
+//       showDenyButton: true,
+//       showCancelButton: true,
+//       confirmButtonText: "Delete",
+//       denyButtonText: `Don't delete`
+//     }).then(async (result) => {
+//       if (result.isConfirmed) {
+//         try {
+//           await deleteBook(id);
+//           fetchBooks();
+//           Swal.fire("Deleted!", "", "success");
+//         } catch (error) {
+//           console.error(error);
+//         }
+//       } else if (result.isDenied) {
+//         Swal.fire("Book is not deleted", "", "info");
+//       }
+//     });
 //   };
 
 //   const handleUpdateBook = async () => {
-//     try {
-//       await updateBook(updateBookId, newBook);
-//       fetchBooks();
-//       setUpdateBookId(null);
-//       setNewBook({ name: '', author: '', genre: '', limit_age: '', status: '' });
-//     } catch (error) {
-//       console.error(error);
-//     }
+//     Swal.fire({
+//       title: "Do you want to update the book?",
+//       showDenyButton: true,
+//       showCancelButton: true,
+//       confirmButtonText: "Update",
+//       denyButtonText: `Don't update`
+//     }).then(async (result) => {
+//       if (result.isConfirmed) {
+//         try {
+//           await updateBook(updateBookId, newBook);
+//           fetchBooks();
+//           setUpdateBookId(null);
+//           setNewBook({ name: '', author: '', genre: '', limit_age: '', status: '' });
+//           Swal.fire("Updated!", "", "success");
+//         } catch (error) {
+//           console.error(error);
+//         }
+//       } else if (result.isDenied) {
+//         Swal.fire("Changes are not saved", "", "info");
+//       }
+//     });
 //   };
 
 //   return (
@@ -74,7 +96,8 @@
 //       <br></br>
 //       <Link to='/demo'>
 //         <button>To the Next Page</button>
-//       </Link>      <div>
+//       </Link>
+//       <div>
 //         <input
 //           type="text"
 //           placeholder="Name"
@@ -129,9 +152,9 @@
 
 // export default Book;
 
-
+// src/Books.js
 import React, { useState, useEffect } from 'react';
-import { getAllBooks, createBook, deleteBook, updateBook } from '../Servers/bookAPI';
+import { getAllBooks, deleteBook, updateBook } from '../Servers/bookAPI';
 import { Link } from 'react-router-dom';
 import Swal from 'sweetalert2';
 
@@ -151,29 +174,6 @@ const Book = () => {
     } catch (error) {
       console.error(error);
     }
-  };
-
-  const handleCreateBook = async () => {
-    Swal.fire({
-      title: "Do you want to add the book?",
-      showDenyButton: true,
-      showCancelButton: true,
-      confirmButtonText: "Save",
-      denyButtonText: `Don't save`
-    }).then(async (result) => {
-      if (result.isConfirmed) {
-        try {
-          await createBook(newBook);
-          fetchBooks();
-          setNewBook({ name: '', author: '', genre: '', limit_age: '', status: '' });
-          Swal.fire("Saved!", "", "success");
-        } catch (error) {
-          console.error(error);
-        }
-      } else if (result.isDenied) {
-        Swal.fire("Changes are not saved", "", "info");
-      }
-    });
   };
 
   const handleDeleteBook = async (id) => {
@@ -225,59 +225,24 @@ const Book = () => {
   return (
     <div>
       <h1>Books</h1>
-      <br></br>
-      <Link to='/demo'>
-        <button>To the Next Page</button>
+      <br />
+      <Link to='/add-book'>
+        <button>Add New Book</button>
       </Link>
       <div>
-        <input
-          type="text"
-          placeholder="Name"
-          value={newBook.name}
-          onChange={(e) => setNewBook({ ...newBook, name: e.target.value })}
-        />
-        <input
-          type="text"
-          placeholder="Author"
-          value={newBook.author}
-          onChange={(e) => setNewBook({ ...newBook, author: e.target.value })}
-        />
-        <input
-          type="text"
-          placeholder="Genre"
-          value={newBook.genre}
-          onChange={(e) => setNewBook({ ...newBook, genre: e.target.value })}
-        />
-        <input
-          type="text"
-          placeholder="Limit Age"
-          value={newBook.limit_age}
-          onChange={(e) => setNewBook({ ...newBook, limit_age: e.target.value })}
-        />
-        <input
-          type="number"
-          placeholder="status"
-          value={newBook.status}
-          onChange={(e) => setNewBook({ ...newBook, status: e.target.value })}
-        />
-        {updateBookId ? (
-          <button onClick={handleUpdateBook}>Update Book</button>
-        ) : (
-          <button onClick={handleCreateBook}>Create Book</button>
-        )}
+        <ul>
+          {books.map((book) => (
+            <li key={book._id}>
+              {book.name} - {book.author}
+              <button onClick={() => handleDeleteBook(book._id)}>Delete</button>
+              <button onClick={() => {
+                setNewBook(book);
+                setUpdateBookId(book._id);
+              }}>Edit</button>
+            </li>
+          ))}
+        </ul>
       </div>
-      <ul>
-        {books.map((book) => (
-          <li key={book._id}>
-            {book.name} - {book.author}
-            <button onClick={() => handleDeleteBook(book._id)}>Delete</button>
-            <button onClick={() => {
-              setNewBook(book);
-              setUpdateBookId(book._id);
-            }}>Edit</button>
-          </li>
-        ))}
-      </ul>
     </div>
   );
 };
